requirements.txt
asgiref==3.5.2
Django==4.0.5
factory-boy==3.2.1
Faker==13.14.0
mysqlclient==2.1.1
Pillow==9.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1

pip==21.3.1
wheel==0.37.1
setuptools==60.2.0
djangorestframework==3.13.1